app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.31.0",
    "json-repair>=0.30.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.9",
    "numpy>=2.2.6",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.3.0",
    "pure-python-adb>=0.3.0.dev0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "openai>=1.40.0",
    "pandas>=2.2.0",
    "pytesseract>=0.3.13",
    "mediapipe>=0.10.33",
    "opencv-python>=4.13.0.92",
    "imagehash>=4.3.2",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.3.0",
    "scikit-learn>=1.8.0",
]
